move switch cam to function

+++ Physics/Source/ChengRigidbody.cpp
@@ -34,7 +34,7 @@ void ChengRigidbody::Update(double dt)
 	Vector3 vAAccel = this->m_vTorque * (1.f / I);
 	m_vAVel += vAAccel * (float)dt;
 	float deg = Trans->GetDegrees();
-	deg += m_vAVel.y * dt;
+	deg += m_vAVel.y * (float)dt;
 	if (m_vAVel.y != 0)
 		Trans->SetRotation(deg, 0, 1, 0);
 	//m_vAVel.SetZero();
